pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "milvus-lite>=2.5.1",
    "openai>=1.93.0",
    "pocketflow>=0.0.1",
    "pydantic>=2.11.7",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]
